+++ lib/boilerpipe/extractors/canola_extractor.rb
@@ -1,0 +1,16 @@
+module Boilerpipe::Extractors
+  class CanolaExtractor
+
+    def self.text(contents)
+      doc = ::Boilerpipe::SAX::BoilerpipeHTMLParser.parse(contents)
+      ::Boilerpipe::Extractors::CanolaExtractor.process doc
+      doc.content
+    end
+
+    def self.process(doc)
+      ::Boilerpipe::Filters::CanolaClassifier.process doc
+
+      doc
+    end
+  end
+end