+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add LargestContent Extractor
 * Add KeepEverything Extractor
 * Add NumWordsRules Extractor
+* Add Canola Extractor
 
 # 0.2.0 / 2017-09-11